# backend/api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8443
  description: Generated server url
paths:
...
  /api/v1/products/{id}:
    get:
      tags:
      - product-rest-controller
      operationId: getProductById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
    put:
      tags:
      - product-rest-controller
      operationId: updateProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdateRequestDTO"
        required: true
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
    delete:
      tags:
      - product-rest-controller
      operationId: deleteProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
...
  /api/v1/admin/users/{id}:
    put:
      tags:
      - admin-rest-controller
      operationId: updateUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdateRequestDTO"
        required: true
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    delete:
      tags:
      - admin-rest-controller
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/admin/categories/{id}:
    put:
      tags:
      - admin-rest-controller
      operationId: updateCategory
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCategoryRequestDTO"
        required: true
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    delete:
      tags:
      - admin-rest-controller
      operationId: deleteCategory
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/v1/admin/categories:
    get:
      tags:
      - admin-rest-controller
      operationId: getCategories
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    post:
      tags:
      - admin-rest-controller
      operationId: createCategory
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCategoryRequestDTO"
        required: true
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/admin/users/{id}/unban:
    patch:
      tags:
      - admin-rest-controller
      operationId: unbanUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/admin/users/{id}/ban:
    patch:
      tags:
      - admin-rest-controller
      operationId: banUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/v1/home/products:
    get:
      tags:
      - home-rest-controller
      operationId: products
      parameters:
      - name: categoryId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: q
        in: query
        required: false
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 1
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/HomeProductsResponseDTO"
...
  /api/v1/products/{productId}/images/{imageId}:
    delete:
      tags:
      - product-rest-controller
      operationId: deleteProductImage
      parameters:
      - name: productId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: imageId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ImageDTO"
components:
  schemas:
...
    ProductUpdateRequestDTO:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
          format: double
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        categoryId:
          type: integer
          format: int64
    CategoryDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        slug:
          type: string
        icon:
          type: string
    ImageDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        url:
          type: string
    ProductDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        price:
          type: number
          format: double
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        averageRating:
          type: number
          format: double
        category:
          $ref: "#/components/schemas/CategoryDTO"
        images:
          type: array
          items:
            $ref: "#/components/schemas/ImageDTO"
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/ProductReviewDTO"
    ProductReviewDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        content:
          type: string
        rating:
          type: number
          format: double
        date:
          type: string
          format: date
        authorName:
          type: string
    AdminUserUpdateRequestDTO:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        fullName:
          type: string
        birthDate:
          type: string
        shippingAddress:
          type: string
    AdminCategoryRequestDTO:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
...
    HomeProductsResponseDTO:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/ProductDTO"
        hasMore:
          type: boolean
...