# backend/src/main/java/es/urjc/daw04/model/dto/ProductUpdateRequestDTO.java
package es.urjc.daw04.model.dto;

import java.util.List;

public record ProductUpdateRequestDTO(
        String name,
        Double price,
        String description,
        List<String> tags,
        Long categoryId
) {
}